[mlir] Resolve CodeQL high/critical alerts in MLIR sources cpp/integer-multiplication-cast-to-long (15) — widen operands to 64-bit before the assignment/index so the multiply no longer overflows in 32-bit:   - mlir/lib/Analysis/Presburger/Matrix.cpp                         (4)   - mlir/lib/Dialect/Linalg/Transforms/ConvertConv2DToImg2Col.cpp   (6; declare tensor dims int64_t, matching the file's other conversions)   - mlir/lib/Dialect/GPU/Transforms/SubgroupReduceLowering.cpp      (1)   - mlir/lib/Dialect/Vector/Transforms/LowerVectorTranspose.cpp     (1)   - mlir/lib/Dialect/Vector/Transforms/VectorEmulateNarrowType.cpp  (1 at resize(); the store-path alert was already 64-bit)   - mlir/tools/mlir-tblgen/OpGenHelpers.cpp                         (1) cpp/suspicious-pointer-scaling (3) — false positives on MLIR's intentional reverse-order co-allocated result storage; annotated with // codeql[cpp/suspicious-pointer-scaling] + justification instead of rewriting core pointer arithmetic:   - mlir/include/mlir/IR/Operation.h                                (2)   - mlir/lib/IR/Value.cpp                                           (1)

+++ mlir/lib/Analysis/Presburger/Matrix.cpp
@@ -4,6 +4,9 @@
 // See https://llvm.org/LICENSE.txt for license information.
 // SPDX-License-Identifier: Apache-2.0 WITH LLVM-exception
 //
+// Modifications (c) Copyright 2026 Advanced Micro Devices, Inc. or its
+// affiliates
+//
 //===----------------------------------------------------------------------===//
 
 #include "mlir/Analysis/Presburger/Matrix.h"
@@ -23,8 +26,9 @@ Matrix<T>::Matrix(unsigned rows, unsigned columns, unsigned reservedRows,
                   unsigned reservedColumns)
     : nRows(rows), nColumns(columns),
       nReservedColumns(std::max(nColumns, reservedColumns)),
-      data(nRows * nReservedColumns) {
-  data.reserve(std::max(nRows, reservedRows) * nReservedColumns);
+      data(static_cast<size_t>(nRows) * nReservedColumns) {
+  data.reserve(static_cast<size_t>(std::max(nRows, reservedRows)) *
+               nReservedColumns);
 }
 
 /// We cannot use the default implementation of operator== as it compares
@@ -128,12 +132,12 @@ void Matrix<T>::swapColumns(unsigned column, unsigned otherColumn) {
 
 template <typename T>
 MutableArrayRef<T> Matrix<T>::getRow(unsigned row) {
-  return {&data[row * nReservedColumns], nColumns};
+  return {&data[static_cast<size_t>(row) * nReservedColumns], nColumns};
 }
 
 template <typename T>
 ArrayRef<T> Matrix<T>::getRow(unsigned row) const {
-  return {&data[row * nReservedColumns], nColumns};
+  return {&data[static_cast<size_t>(row) * nReservedColumns], nColumns};
 }
 
 template <typename T>
